e2e: start a new file for cli plugins arguments tests
`TestRunGoodArgument` fits here.
